+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.26.0
 
+- **Fix** : Lock order table row while executing notification
 - **Add** : Dynamically get Alma MAX/MIN from merchant account
 - **Fix** : Lock order table row while executing notification
 - **Fix** : Allow retrieving customer X-Forwarded-For IP addresse